fix/feat!(main.rs): correctly resolve symlinks
Closes: #339

Addresses issue where executable name 
resolution fails when symlinked wtih a 
name that is not `nvim`.

With this change a link from
`$HOME/.local/share/bob/nvim-bin/nvim`
to any name will still resolve as intended.

When the executable was run via a symlink,
the previous method of
extracting would only look at the output 
side of the symlink, leading to incorrect 
behavior.